requirements.txt
# =============================================================================
# MSAI Projects - Consolidated Requirements
# =============================================================================
# Version: 1.1.0
# Last Updated: 2025-12-10
# Updated for: Python 3.11 + PyTorch 2.7.1 + TorchAO 0.14.1
#
# Projects supported:
#   - UdaciSense (Model Optimization)
#   - GPT-2 Model Optimization
#   - CharityML (Finding Donors)
#   - Image Classification (CIFAR-10)
#   - Customer Segmentation
#   - General ML/AI projects
#
# Installation Instructions:
#   1. Create conda environment: conda create -n msai python=3.11
#   2. Activate environment: conda activate msai
#   3. Install PyTorch 2.7.1 with CUDA 12.8 support (optimized for RTX 3090):
#      pip install torch==2.7.1 torchvision==0.22.1 torchaudio==2.7.1 --index-url https://download.pytorch.org/whl/cu128
#   4. Install TorchAO for advanced quantization:
#      pip install torchao==0.12.1
#   5. Install remaining packages: pip install -r requirements.txt
#
# Note: This setup provides stable quantization support with TorchAO.
#       PyTorch 2.7.1 + Python 3.11 is the recommended stable combination.
# =============================================================================

# ==========================
# PyTorch Ecosystem (Install separately with CUDA support)
# ==========================
# torch==2.7.1
# torchvision==0.22.1
# torchaudio==2.7.1

# ==========================
# Core Data Science
# ==========================
numpy>=1.26.0,<2.0.0
pandas>=2.2.0
scipy>=1.14.0
scikit-learn>=1.5.0

# ==========================
# Visualization
# ==========================
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.24.0
bokeh>=3.4.0

# ==========================
# Jupyter & Interactive Computing
# ==========================
jupyter>=1.0.0
jupyterlab>=4.2.0
notebook>=7.2.0
ipykernel>=6.29.0
ipywidgets>=8.1.0
ipython>=8.20.0

# ==========================
# Deep Learning & Model Optimization
# ==========================
tensorboard>=2.18.0
thop>=0.1.1  # FLOPs calculation

# ==========================
# Hugging Face Ecosystem
# ==========================
transformers>=4.45.0
datasets>=3.0.0
huggingface_hub>=0.25.0
accelerate>=1.0.0
evaluate>=0.4.0

# ==========================
# Model Compression & Quantization
# ==========================
torchao==0.12.0  # Modern PyTorch quantization library (compatible with PyTorch 2.7.1)
optimum[neural-compressor]>=1.23.0
neural-compressor>=3.0

# ==========================
# Machine Learning Frameworks
# ==========================
xgboost>=2.0.3
#autogluon>=1.1.0

# ==========================
# Computer Vision
# ==========================
pillow>=10.2.0

# ==========================
# NLP & Text Processing
# ==========================
nltk>=3.8.1
beautifulsoup4>=4.12.3

# ==========================
# Data Handling & Cloud
# ==========================
boto3>=1.34.34
requests>=2.31.0
pyarrow>=15.0.0

# ==========================
# Utilities
# ==========================
tqdm>=4.66.0
psutil>=6.0.0
python-dotenv>=1.0.0
markdown>=3.3.0
Jinja2>=3.1.3

# ==========================
# MCP / PriceScout (Competitor Analyst Bot)
# ==========================
aiohttp>=3.12.15
anthropic>=0.64.0
firecrawl-py>=2.16.5
mcp>=1.10.1
mcp-server-sqlite

# ==========================
# Optional: Performance Profiling
# ==========================
py-spy>=0.3.14

# ==========================
# Platform-Specific (Conditional)
# ==========================
# MXNet (Linux/Mac only, for CharityML)
mxnet>=1.9.1; sys_platform != 'win32'
